+++ CHANGELOG.md
@@ -44,6 +44,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - fix checkbox questions where further information couldn't be saved for an option that included a special character
 - further information for options that include special characters is now available to the specification
 - incomplete specification documents include a warning within the document as well as the file name
+- users return to the same place in the task list after answering a question
 
 ## [release-005] - 2021-1-19